+++ base/oix/fmonitr.r
@@ -323,8 +323,8 @@ void EVVariable(dptr dx, int eventcode)
 
    if (i == GlobalName) {
       if (reserve(Strings, StrLen(parent->eventval) + 1) == NULL) {
-	 fprintf(stderr, "failed to reserve %d bytes for global\n",
-		 StrLen(parent->eventval)+1);
+	 fprintf(stderr, "failed to reserve %ld bytes for global\n",
+		 (long)StrLen(parent->eventval)+1);
 	 syserr("monitoring out-of-memory error");
 	 }
       StrLoc(parent->eventval) =
@@ -334,9 +334,9 @@ void EVVariable(dptr dx, int eventcode)
       }
    else if ((i == StaticName) || (i == LocalName) || (i == ParamName)) {
       if (!reserve(Strings, StrLen(parent->eventval) + StrLen(*procname) + 1)) {
-	 fprintf(stderr,"failed to reserve %d bytes for %d, %d+%d\n",
-		StrLen(parent->eventval)+StrLen(*procname)+1, i,
-		 StrLen(parent->eventval), StrLen(*procname));
+	 fprintf(stderr,"failed to reserve %ld bytes for %d, %ld+%ld\n",
+                 (long)StrLen(parent->eventval)+(long)StrLen(*procname)+1, i,
+		 (long)StrLen(parent->eventval), (long)StrLen(*procname));
 	 syserr("monitoring out-of-memory error");
 	 }
       StrLoc(parent->eventval) =
